Melhorias

+++ Desafio01_Conceitos/Exercicios04_conceito.py
@@ -3,8 +3,8 @@ Receba do usuário a quantidade de litros de combustível consumidos e
 a distância percorrida. Calcule e imprima o consumo médio em km/l.
 '''
 
-litros = int(input('Digite a quantidade de litros de combustível: '))
-distancia = int(input('Digite a distância percorrida em km: '))
+litros = float(input('Digite a quantidade de litros de combustível: '))
+distancia = float(input('Digite a distância percorrida em km: '))
 
 consumo = distancia/litros
